+++ GenerateTestJSON.hs
@@ -3,6 +3,7 @@
 {-# LANGUAGE OverloadedStrings    #-}
 {-# LANGUAGE DeriveGeneric        #-}
 {-# LANGUAGE StandaloneDeriving   #-}
+{-# LANGUAGE RecordWildCards      #-}
 {-# LANGUAGE ViewPatterns         #-}
 {-# LANGUAGE FlexibleInstances    #-}
 {-# LANGUAGE FlexibleContexts     #-}
@@ -21,6 +22,7 @@ import           Control.Monad             (forM_, forM, when)
 import qualified Data.ByteString.Lazy.Char8 as BSL
 import qualified Data.HashMap.Strict        as Map
 import qualified Data.Set                   as Set
+import           Data.Semigroup            ((<>))
 import           Data.Aeson
 import           Data.Function             (on)
 import           Data.List
@@ -39,24 +41,56 @@ import           Data.Aeson.AutoType.Format
 import           Data.Aeson.AutoType.CodeGen
 import           Data.Aeson.AutoType.Util
 import           Data.Aeson.AutoType.Test
-import           HFlags
+--import           HFlags -- TODO: remove
+import           Options.Applicative
 
 -- * Command line flags
-defineFlag "z:size"            (10     :: Int)        "Size of generated elements"
-defineFlag "s:stem"            ("Test" :: FilePath)   "Test filename stem"
-defineFlag "c:count"           (100    :: Int)        "Number of test cases to generate."
+--defineFlag "z:size"            (10     :: Int)        "Size of generated elements"
+--defineFlag "s:stem"            ("Test" :: FilePath)   "Test filename stem"
+--defineFlag "c:count"           (100    :: Int)        "Number of test cases to generate."
 --defineFlag "o:outputFilename"   defaultOutputFilename "Write output to the given file"
-flags_suggest = True
+--flags_suggest = True
 --defineFlag "suggest"            True                  "Suggest candidates for unification"
-defineFlag "autounify"          True                  "Automatically unify suggested candidates"
-defineFlag "t:test"             True                  "Try to run generated parser after"
-defineFlag "d:debug"            False                 "Set this flag to see more debugging info"
-defineFlag "keep"               False                 "Keep also the successful tests"
-defineFlag "fakeFlag"           True                  "Ignore this flag - it doesn't exist!!! It is workaround for a library problem."
+--defineFlag "autounify"          True                  "Automatically unify suggested candidates"
+--defineFlag "t:test"             True                  "Try to run generated parser after"
+--defineFlag "d:debug"            False                 "Set this flag to see more debugging info"
+--defineFlag "keep"               False                 "Keep also the successful tests"
+--defineFlag "fakeFlag"           True                  "Ignore this flag - it doesn't exist!!! It is workaround for a library problem."
 
--- Tracing is switched off:
-myTrace :: String -> IO ()
-myTrace msg = flags_debug `when` putStrLn msg
+data TypeOpts = TyOptions {
+                  autounify :: Bool
+                , debug     :: Bool
+                , test      :: Bool
+                , suggest   :: Bool
+                }
+
+data Options = Options {
+                 tyOpts    :: TypeOpts
+               , keep      :: Bool
+               , stem      :: FilePath
+               , count     :: Int
+               , size      :: Int
+               }
+
+unflag = flag True False
+
+tyOptParser :: Parser TypeOpts
+tyOptParser  = TyOptions
+            <$> unflag (long "no-autounify" <> help "Do not automatically unify suggested candidates")
+            <*> switch (long "debug"        <> help "Set this flag to see more debugging info"       )
+            <*> unflag (long "no-test"      <> help "Do not run generated parser afterwards"         )
+            <*> unflag (long "no-suggest"   <> help "Do not suggest candidates for unification"      )
+
+optParser :: Parser Options
+optParser  =
+    Options  <$> tyOptParser
+             <*> switch    (long "keep"                  <> help "Also keep successful tests"  )
+             <*> strOption (long "stem"  <> value "Test" <> help "Output filename stem"        )
+             <*> intOpt    (long "count" <> value 100    <> help "Number of tests to perform"  )
+             <*> intOpt    (long "size"  <> value 10     <> help "size of generated test cases")
+             -- <*> some (argument str (metavar "FILES..."))
+  where
+    intOpt = option auto
 
 -- | Report an error to error output.
 report   :: Text -> IO ()
@@ -68,8 +102,8 @@ fatal msg = do report msg
                exitFailure
 
 -- | Read JSON and extract @Type@ information from it.
-extractTypeFromJSONFile :: FilePath -> IO (Maybe Type)
-extractTypeFromJSONFile inputFilename =
+extractTypeFromJSONFile :: (String -> IO ()) -> FilePath -> IO (Maybe Type)
+extractTypeFromJSONFile myTrace inputFilename =
       withFileOrHandle inputFilename ReadMode stdin $ \hIn ->
         -- First we decode JSON input into Aeson's Value type
         do bs <- BSL.hGetContents hIn
@@ -83,6 +117,7 @@ extractTypeFromJSONFile inputFilename =
                let t        = extractType v
                myTrace $ "Type: " ++ pretty t
                return $ Just t
+
 
 vectorWithoutDuplicates ::  Ord b => Int -> Gen b -> Gen [b]
 vectorWithoutDuplicates i gen = take i
@@ -120,16 +155,16 @@ instance Ord Value where
   (Object a) `compare` (Object b) = Map.toList a `compare` Map.toList b
 
 -- | Take a set of JSON input filenames, Haskell output filename, and generate module parsing these JSON files.
-generateTestJSONs :: IO ()
-generateTestJSONs = do
+generateTestJSONs :: Options -> IO ()
+generateTestJSONs Options {tyOpts=TyOptions {..}, ..}= do
     testValues :: [Value] <- generate $
-                               resize flags_size $
+                               resize size $
                                  vectorWithoutDuplicates 100 arbitraryTopValue
     results               <- forM (zip3 inputFilenames outputFilenames testValues) $
       \(inputFilename, outputFilename, jsonValue) -> do
         BSL.writeFile inputFilename $ encode jsonValue
         -- Read type from each file
-        typeForEachFile  <- catMaybes <$> mapM extractTypeFromJSONFile [inputFilename]
+        typeForEachFile  <- catMaybes <$> mapM (extractTypeFromJSONFile myTrace) [inputFilename]
         -- Unify all input types
         when (null typeForEachFile) $ do
           report "No valid JSON input file..."
@@ -142,17 +177,17 @@ generateTestJSONs = do
         -- We compute which type labels are candidates for unification
         let uCands = unificationCandidates splitted
         myTrace $ "CANDIDATES:\n" ++ pretty uCands
-        when flags_suggest $ forM_ uCands $ \cs -> do
+        when suggest $ forM_ uCands $ \cs -> do
                                putStr "-- "
                                Text.putStrLn $ "=" `Text.intercalate` cs
         -- We unify the all candidates or only those that have been given as command-line flags.
-        let unified = if flags_autounify
+        let unified = if autounify
                         then unifyCandidates uCands splitted
                         else splitted
         myTrace $ "UNIFIED:\n" ++ pretty unified
         -- We start by writing module header
         writeHaskellModule outputFilename unified
-        if flags_test
+        if test
           then do
             r <- (==ExitSuccess) <$> system (unwords ["runghc", "-package=aeson-0.9.0.1", outputFilename, inputFilename])
             when r $ mapM_ removeFile [inputFilename, outputFilename]
@@ -160,15 +195,20 @@ generateTestJSONs = do
           else
             return True
     putStrLn $ "Successfully generated "      ++ show (length results) ++
-               " JSON files, out of planned " ++ show flags_count  ++ " cases."
+               " JSON files, out of planned " ++ show count  ++ " cases."
   where
-    makeInputFilename  = (<.>".json") . (flags_stem ++) . show
-    makeOutputFilename = (<.>".hs")   . (flags_stem ++) . show
-    inputFilenames     = map makeInputFilename  [1..flags_count]
-    outputFilenames    = map makeOutputFilename [1..flags_count]
+    makeInputFilename  = (<.>".json") . (stem ++) . show
+    makeOutputFilename = (<.>".hs")   . (stem ++) . show
+    inputFilenames     = map makeInputFilename  [1..count]
+    outputFilenames    = map makeOutputFilename [1..count]
+    myTrace :: String -> IO ()
+    myTrace msg = debug `when` putStrLn msg
 
 main :: IO ()
-main = do filenames <- $initHFlags "testJSON -- automatic test JSON generation"
-          -- TODO: should integrate all inputs into single type set!!!
-          generateTestJSONs
-
+main = do opts <- execParser optInfo
+          generateTestJSONs opts
+    where
+      optInfo = info (optParser <**> helper)
+        ( fullDesc
+       <> progDesc "Generate a number of JSON test files, and generate type and parser for each."
+       <> header   "Self-test for json-autotype" )